backend/requirements.txt
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.103.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
Mako==1.2.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
